@@ -1,14 +1,12 @@
 <template>
-  <div class="section-home row" :style="{ backgroundImage: `url(${backgroundImageProfil})` }">
+  <div class="section-home row">
     <div
-      class="col-xs-12 col-md-6 flex flex-center"
-      :style="{ backgroundImage: `url(${backgroundImageProfil2})` }"
+      class="section-home-profil-container col-xs-12 col-md-6 flex flex-center"
     >
       <section-home-profil />
     </div>
     <div
-      class="col-xs-12 col-md-6 flex flex-center"
-      :style="{ backgroundImage: `url(${backgroundImageProfil2})` }"
+      class="section-home-profil-container col-xs-12 col-md-6 flex flex-center"
     >
       <section-home-welcome ref="homeWelcome" />
     </div>
@@ -17,6 +15,7 @@
 
 <script setup lang="ts">
 import { ref } from 'vue';
+
 import SectionHomeWelcome from '@/components/sectionHome/sectionHomeWelcome/SectionHomeWelcome.vue';
 import SectionHomeProfil from '@/components/sectionHome/SectionHomeProfil.vue';
 
@@ -37,25 +36,36 @@ useMotion(homeWelcome, {
     },
   },
 });
-
-const backgroundImageProfil = new URL(
-  'http://www.ivang-design.com/svg-load/parallax/parallax-01.png',
-  import.meta.url
-);
-const backgroundImageProfil2 = new URL(
-  'http://www.ivang-design.com/svg-load/parallax/parallax-0.png',
-  import.meta.url
-);
 </script>
 
 <style lang="scss">
 .section-home {
+  background-image: url(https://erickgolos.com/images/home/background-1.png);
+
   @media (min-width: $breakpoint-sm-min) {
     height: auto;
   }
 
   @media (min-width: $breakpoint-md-min) {
     height: calc(100vh - 50px) !important;
+  }
+
+  .section-home-profil-container {
+    background-image: url(https://erickgolos.com/images/home/background-0.png);
+  }
+
+  .section-home-profil-container {
+    background-image: url(https://erickgolos.com/images/home/background-0.png);
+  }
+
+  .gradient-left {
+    left: 0;
+    top: 0;
+  }
+
+  .gradient-right {
+    right: 0;
+    top: 0;
   }
 }
 </style>
