@@ -1,24 +1,21 @@
 <template>
-  <q-parallax class="picture-profil q-pa-md q-mt-md" :src="backgroundProfil">
+  <q-parallax class="picture-profil q-pa-md q-mt-md">
     <q-img
       class="picture-profil bg-transparent shadow-10 border-primary"
-      :src="pictureProfil"
+      src="https://erickgolos.com/images/home/photo.png"
       fit="contain"
     />
   </q-parallax>
 </template>
 
-<script setup lang="ts">
-import pictureProfil from '@/assets/general/photo.png';
-import backgroundProfil from '@/assets/general/mountain.jpg';
-</script>
+<script setup lang="ts"></script>
 
 <style lang="scss">
 @use '@/css/quasar.variables.scss';
 
 .picture-profil {
-  width: 30vh;
-  height: 30vh !important;
+  width: 25vh;
+  height: 25vh !important;
   border-radius: 50%;
 
   @media (min-width: $breakpoint-sm-min) {
@@ -35,10 +32,10 @@ import backgroundProfil from '@/assets/general/mountain.jpg';
   }
 
   &.border-primary {
-    border: solid 2px $primary;
+    border: solid 2px var(--q-primary);
 
     @media (min-width: $breakpoint-sm-min) {
-      border: solid 10px $primary;
+      border: solid 10px var(--q-primary);
     }
   }
 }
