@@ -1,8 +1,14 @@
 <template>
   <div class="q-gutter-x-md">
-    <q-btn round color="indigo-7" icon="eva-linkedin" />
-    <q-btn round color="grey-8" icon="eva-github" />
-    <q-btn round color="red-8" icon="eva-email" />
+    <q-btn
+      round
+      color="indigo-7"
+      icon="eva-linkedin"
+      href="https://www.linkedin.com/in/erick-golos-ingenieur-informatique/"
+      target="_blank"
+    />
+    <q-btn round color="grey-8" icon="eva-github" href="https://github.com/Lurrik" target="_blank" />
+    <q-btn round color="red-8" icon="eva-email" href="mailto:erickgolos@gmail.com" />
   </div>
 </template>
 
