@@ -1,40 +1,47 @@
 <template>
   <q-tabs
     v-model="tab"
-    class="the-navigation"
+    class="the-header-navigation"
     :breakpoint="500"
     active-color="primary"
     no-caps
     indicator-color="transparent"
     inline-label
   >
-    <q-tab name="home" icon="fa-solid fa-house-user" label="Accueil" />
-    <q-tab name="about" icon="fa-solid fa-user-astronaut" label="À propos de moi" />
-    <q-tab name="skills" icon="fa-solid fa-code" label="Expertises" />
-    <q-tab name="experience" icon="fa-solid fa-timeline" label="Expériences" />
-    <q-tab name="projects" icon="fa-solid fa-list-check" label="Projets" />
-    <q-tab name="contact" icon="fa-solid fa-envelope" label="Contact" />
+    <q-route-tab
+      name="home"
+      to="#home"
+      exact
+      replace
+      icon="fa-solid fa-house-user"
+      label="Accueil"
+    />
+    <q-route-tab
+      name="skills"
+      to="#skills"
+      exact
+      replace
+      icon="fa-solid fa-code"
+      label="Expertises"
+    />
+    <q-route-tab
+      name="projects"
+      to="#projects"
+      exact
+      replace
+      icon="fa-solid fa-list-check"
+      label="Projets"
+    />
   </q-tabs>
 </template>
 
 <script setup lang="ts">
-import { computed } from 'vue';
-import { useNavigationStore } from '@/stores/useNavigationStore';
-
-const store = useNavigationStore();
-
-const tab = computed<string>({
-  get() {
-    return store.navigation;
-  },
-  set(value: string) {
-    store.setNavigation(value);
-  },
-});
+import { ref } from 'vue';
+const tab = ref('home');
 </script>
 
 <style lang="scss">
-.the-navigation {
+.the-header-navigation {
   .q-tab__icon {
     width: 16px;
     height: 16px;
